Fix existing tests
