+++ kafka-impl/src/main/java/io/streamnative/pulsar/handlers/kop/InternalServerCnx.java
@@ -16,7 +16,6 @@ package io.streamnative.pulsar.handlers.kop;
 import java.net.InetSocketAddress;
 import java.net.SocketAddress;
 import lombok.Getter;
-import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.pulsar.broker.service.Producer;
 import org.apache.pulsar.broker.service.ServerCnx;
@@ -29,6 +28,9 @@ import org.apache.pulsar.broker.service.ServerCnx;
  */
 @Slf4j
 public class InternalServerCnx extends ServerCnx {
+
+    public static final SocketAddress MOCKED_REMOTE_ADDRESS = new InetSocketAddress("localhost", 9999);
+
     @Getter
     KafkaRequestHandler kafkaRequestHandler;
 
@@ -41,7 +43,7 @@ public class InternalServerCnx extends ServerCnx {
         // mock some values, or Producer create will meet NPE.
         // used in test, which will not call channel.active, and not call updateCtx.
         if (this.remoteAddress == null) {
-            this.remoteAddress = new InetSocketAddress("localhost", 9999);
+            this.remoteAddress = MOCKED_REMOTE_ADDRESS;
         }
     }
 
@@ -58,8 +60,8 @@ public class InternalServerCnx extends ServerCnx {
     }
 
     // called after channel active
-    public void updateCtx() {
-        this.remoteAddress = kafkaRequestHandler.remoteAddress;
+    public void updateCtx(final SocketAddress remoteAddress) {
+        this.remoteAddress = remoteAddress;
     }
 
     @Override
@@ -75,9 +77,5 @@ public class InternalServerCnx extends ServerCnx {
     @Override
     public void cancelPublishBufferLimiting() {
         // do nothing is this mock
-    }
-
-    public @NonNull SocketAddress getRemoteAddress() {
-        return remoteAddress;
     }
 }
